refactor(git): GitDiffPanel 续拆 contextMenus 与 preview 两段（2711→2146）
